backend/requirements.txt
asgiref==3.5.2
azure-core==1.24.1
azure-storage-blob==12.12.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.5
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
msrest==0.7.0
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
gunicorn==20.1.0